extract from test function needed in normal code
