+++ vispy/testing/__init__.py
@@ -42,8 +42,8 @@ to determine what automatic tests are run.
 
 from ._testing import (SkipTest, requires_application, requires_ipython,  # noqa
                        requires_img_lib,  # noqa
-                       has_backend, requires_pyopengl,  # noqa
-                       requires_scipy, has_matplotlib, has_skimage,  # noqa
+                       requires_pyopengl,  # noqa
+                       requires_scipy,  # noqa
                        save_testing_image, TestingCanvas, has_pyopengl,  # noqa
                        run_tests_if_main, requires_ssl,  # noqa
                        assert_is, assert_in, assert_not_in, assert_equal,
