+++ vispy/testing/_testing.py
@@ -17,7 +17,7 @@ import functools
 from distutils.version import LooseVersion
 
 from ..ext.six import string_types
-from ..util import use_log_level
+from ..util.check_environment import has_backend
 
 skipif = pytest.mark.skipif
 
@@ -168,31 +168,6 @@ def requires_ssl():
 ###############################################################################
 # App stuff
 
-def has_backend(backend, has=(), capable=(), out=()):
-    from ..app.backends import BACKENDMAP
-    using = os.getenv('_VISPY_TESTING_APP', None)
-    if using is not None and using != backend:
-        # e.g., we are on  a 'pyglet' run but the test requires PyQt4
-        ret = (False,) if len(out) > 0 else False
-        for o in out:
-            ret += (None,)
-        return ret
-
-    # let's follow the standard code path
-    module_name = BACKENDMAP[backend.lower()][1]
-    with use_log_level('warning', print_msg=False):
-        mod = __import__('app.backends.%s' % module_name, globals(), level=2)
-    mod = getattr(mod.backends, module_name)
-    good = mod.testable
-    for h in has:
-        good = (good and getattr(mod, 'has_%s' % h))
-    for cap in capable:
-        good = (good and mod.capability[cap])
-    ret = (good,) if len(out) > 0 else good
-    for o in out:
-        ret += (getattr(mod, o),)
-    return ret
-
 
 def has_application(backend=None, has=(), capable=()):
     """Determine if a suitable app backend exists"""
@@ -298,30 +273,6 @@ def requires_numpydoc():
     else:
         present = True
     return skipif(not present, reason='numpydoc is required')
-
-
-def has_matplotlib(version='1.2'):
-    """Determine if mpl is a usable version"""
-    try:
-        import matplotlib
-    except Exception:
-        has_mpl = False
-    else:
-        if LooseVersion(matplotlib.__version__) >= LooseVersion(version):
-            has_mpl = True
-        else:
-            has_mpl = False
-    return has_mpl
-
-
-def has_skimage(version='0.11'):
-    """Determine if scikit-image is a usable version"""
-    try:
-        import skimage
-    except ImportError:
-        return False
-    sk_version = LooseVersion(skimage.__version__)
-    return sk_version >= LooseVersion(version)
 
 
 ###############################################################################
